run-onnx/pyproject.toml
[project]
name = "llama-pajamas-run-onnx"
version = "0.1.0"
description = "ONNX Runtime backend for edge deployment (CPU, AMD, ARM, Jetson)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0",
    "pillow>=10.0.0",
    "soundfile>=0.12.1",
    "librosa>=0.10.0",
    "openai-whisper>=20230314",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.16.0",
]
full = [
    "llama-pajamas-run-onnx[gpu]",
    "onnx>=1.15.0",
    "onnxoptimizer>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
